+++ src/zones/cc-zone-input.js
@@ -7,11 +7,10 @@ import { repeat } from 'lit-html/directives/repeat.js';
 import { scrollChildIntoParent } from '../lib/dom.js';
 import { dispatchCustomEvent } from '../lib/events.js';
 import { i18n } from '../lib/i18n.js';
+import { PRIVATE_ZONE, sortZones } from '../lib/zone.js';
 import { withResizeObserver } from '../mixins/with-resize-observer.js';
 
 const SKELETON_ZONES = new Array(6).fill(null);
-const CLEVER_CLOUD_ZONE = 'infra:clever-cloud';
-const PRIVATE_ZONE = 'scope:private';
 
 /**
  * A input component to select a zone with a map and a list.
@@ -56,6 +55,7 @@ export class CcZoneInput extends withResizeObserver(LitElement) {
       _hovered: { type: String },
       _legend: { type: String },
       _points: { type: Array },
+      _sortedZones: { type: Array },
     };
   }
 
@@ -72,12 +72,12 @@ export class CcZoneInput extends withResizeObserver(LitElement) {
 
   _updatePoints () {
 
-    if (!Array.isArray(this.zones)) {
+    if (!Array.isArray(this._sortedZones)) {
       return;
     }
 
     // Filter out private zones from the map for now as they may be at the same coordinates and overlap
-    this._points = this.zones
+    this._points = this._sortedZones
       .filter((zone) => !zone.tags.includes(PRIVATE_ZONE))
       .map((zone) => ({
         name: zone.name,
@@ -88,7 +88,7 @@ export class CcZoneInput extends withResizeObserver(LitElement) {
         zIndexOffset: this._getZIndexOffset(zone.name),
       }));
 
-    this._legend = this.zones.some((zone) => zone.tags.includes(PRIVATE_ZONE))
+    this._legend = this._sortedZones.some((zone) => zone.tags.includes(PRIVATE_ZONE))
       ? i18n('cc-zone-input.private-map-warning')
       : '';
   }
@@ -113,36 +113,6 @@ export class CcZoneInput extends withResizeObserver(LitElement) {
     return 0;
   }
 
-  // 1. Clever Cloud zones "infra:clever-cloud"
-  // 2. Private zones "scope:private"
-  // 3. Alphanum sort on city
-  _sortZones (zones) {
-    return [...zones].sort((a, b) => {
-      if (a == null || b == null) {
-        return 0;
-      }
-      const aIsCcZone = a.tags.includes(CLEVER_CLOUD_ZONE);
-      const bIsCcZone = b.tags.includes(CLEVER_CLOUD_ZONE);
-      if (aIsCcZone !== bIsCcZone) {
-        return aIsCcZone ? -1 : 1;
-      }
-      const aIsPrivateZone = a.tags.includes(PRIVATE_ZONE);
-      const bIsPrivateZone = b.tags.includes(PRIVATE_ZONE);
-      if (aIsCcZone && bIsCcZone) {
-        if (aIsPrivateZone !== bIsPrivateZone) {
-          return aIsPrivateZone ? 1 : -1;
-        }
-        if (aIsPrivateZone && bIsPrivateZone) {
-          return (a.displayName || '').localeCompare((b.displayName || ''));
-        }
-      }
-      if (aIsPrivateZone !== bIsPrivateZone) {
-        return aIsPrivateZone ? -1 : 1;
-      }
-      return a.city.localeCompare(b.city);
-    });
-  }
-
   _onSelect (name) {
     this.selected = name;
     dispatchCustomEvent(this, 'input', this.selected);
@@ -164,11 +134,11 @@ export class CcZoneInput extends withResizeObserver(LitElement) {
     clearTimeout(this._panMapTimeout);
     this._panMapTimeout = setTimeout(() => {
       if (this._hovered != null) {
-        const zone = this.zones.find((z) => z.name === this._hovered);
+        const zone = this._sortedZones.find((z) => z.name === this._hovered);
         this._map.panInside(zone.lat, zone.lon);
       }
       else if (this.selected != null) {
-        const zone = this.zones.find((z) => z.name === this.selected);
+        const zone = this._sortedZones.find((z) => z.name === this.selected);
         this._map.panInside(zone.lat, zone.lon);
       }
     }, 200);
@@ -230,6 +200,7 @@ export class CcZoneInput extends withResizeObserver(LitElement) {
     }
 
     if (changedProperties.has('zones')) {
+      this._sortedZones = sortZones(this.zones);
       this._updatePoints();
     }
 
@@ -238,8 +209,8 @@ export class CcZoneInput extends withResizeObserver(LitElement) {
 
   render () {
 
-    const skeleton = (this.zones == null);
-    const zones = skeleton ? SKELETON_ZONES : this.zones;
+    const skeleton = (this._sortedZones == null);
+    const zones = skeleton ? SKELETON_ZONES : this._sortedZones;
 
     // Try to zoom out and center the map
     return html`