+++ CHANGELOG.md
@@ -18,6 +18,16 @@ title: Changelog
 * `<cc-pricing-table>`: rename `items` to `plans` (BREAKING CHANGE)
 * `<cc-pricing-product>`: rename `items` to `plans` (BREAKING CHANGE)
 * `<cc-block>`: introduce a ribbon and a noHead property
+* `<cc-zone>`:
+  * add CSS custom properties to customize colors
+  * replace rem with em
+* `<cc-zone-input>`:
+  * fix zone list sort
+  * move `sortZones` out in a lib
+* New components:
+  * `<cc-pricing-header>`
+  * `<cc-pricing-estimation>`
+  * `<cc-pricing-page>` (with smart definition)
 
 ## 6.10.0 (2021-07-08)